Updated the differences between buttons as xinput is fills a short.  Might change in the future but it is something

+++ projects/1.1hello_world/1.1hello_world.c
@@ -9,7 +9,7 @@ void app_specific_init(void)
 
 void app_specific_update(double dt)
 {
-    total_time+=dt;
+    total_time += (float)dt;
     //NL_LOG("Frame: %f, total: %f", dt, total_time);
 
     if (was_button_pressed(NL_GAMEPAD_A))
